Remove call to createDeviceInfoList.
This function takes quite a while to run and can cause major slowdown. The d2xx developer documentation explicitly states that functions which rely on the device info list can be called, createDeviceInfoList must be called. A developer following that guide (such as myself) would be thrown off by all the implicit calls to that function.